+++ guillotina/tests/memcached/test_memcached_driver.py
@@ -1,4 +1,5 @@
 from guillotina.contrib.memcached.driver import MemcachedDriver
+from guillotina.contrib.memcached.driver import safe_key
 from guillotina.utils import resolve_dotted_name
 from unittest import mock
 
@@ -99,3 +100,26 @@ async def test_memcached_ops(memcached_container, guillotina_main, loop):
 
     await driver.finalize()
     assert driver.initialized is False
+
+
+unsafe_keys = ["a" * 255, "foo bar", b"\x130".decode()]
+
+
+@pytest.mark.app_settings(MEMCACHED_SETTINGS)
+@pytest.mark.parametrize("unsafe_key", unsafe_keys)
+async def test_memcached_ops_are_safe_key(memcached_container, guillotina_main, loop, unsafe_key):
+    driver = await resolve_dotted_name("guillotina.contrib.memcached").get_driver()
+    await driver.get(unsafe_key)
+    await driver.set(unsafe_key, b"foo")
+    await driver.delete(unsafe_key)
+    await driver.delete_all([unsafe_key])
+
+
+@pytest.mark.parametrize("unsafe_key", unsafe_keys)
+async def test_safe_key(unsafe_key):
+    key = safe_key(unsafe_key)
+    assert isinstance(key, bytes)
+    assert len(key) < 250
+    for char in key:
+        assert char >= 33
+        assert char <= 126